working on conncetedList
